Create Stack.java
A stack is a special kind of LinkedList. It generally only supports fast access to one side of the list as opposed to both and does not allow for access or removal for elements that are not the top of the stack.

Stack is therefore called a LIFO(Last-in First-out) data structure. That is, the last element added to the stack will be the first element to be removed.