Create Queue.java
Queue is again a special type of LinkedList.  Like a stack, a queue does not support the removal of elements in the middle; it only supports deletion from the head of the queue and insertion at the end of the queue.

It is a FIFO(First-in, First-out) data structure. That is, the first element to be added will be the first element to be removed, and so forth.